4.21 valmis

+++ controllers/blogs.js
@@ -46,8 +46,20 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = decodedToken.id
+    const blog = await Blog.findById(request.params.id)
+    if (blog == null) {
+      return response.status(404).json({ error: 'not found' })
+    }
+    if (user != blog.user) {
+      return response.status(401).json({ error: 'not authorized to delete' })
+    }
     const deletedBlog = await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end();
+    response.status(204).end()
   } catch (exception) {
     next(exception)
   }
